Further test fixes and added julia and js metadata

+++ tests/unit/test_artifact.py
@@ -13,12 +13,16 @@ def test_artifact_creation_online() -> None:
     _folder_name = f"/simvue_unit_testing/{_uuid}"
     _folder = Folder.new(path=_folder_name)
     _run = Run.new(folder=_folder_name) 
+    _folder.commit()
+    _run.commit()
+
+    _failed = []
 
     with tempfile.NamedTemporaryFile(suffix=".txt") as temp_f:
         _path = pathlib.Path(temp_f.name)
         with _path.open("w") as out_f:
             out_f.write("Hello World!")
-        _artifact = Artifact.new(
+        _artifact = Artifact.new_file(
             name=f"test_artifact_{_uuid}",
             run=_run.id,
             file_path=_path,
@@ -26,10 +30,16 @@ def test_artifact_creation_online() -> None:
             storage=None,
             file_type=None
         )
-        _artifact.commit()
         time.sleep(1)
+        for member in _artifact._properties:
+            try:
+                getattr(_artifact, member)
+            except Exception as e:
+                _failed.append((member, f"{e}"))
         assert _artifact.name == f"test_artifact_{_uuid}"
         _artifact.delete()
+    if _failed:
+        raise AssertionError("\n" + "\n\t- ".join(": ".join(i) for i in _failed))
     _run.delete()
     _folder.delete()
 
@@ -45,7 +55,7 @@ def test_artifact_creation_offline() -> None:
         _path = pathlib.Path(temp_f.name)
         with _path.open("w") as out_f:
             out_f.write("Hello World!")
-        _artifact = Artifact.new(
+        _artifact = Artifact.new_file(
             name=f"test_artifact_{_uuid}",
             run=_run.id,
             file_path=_path,
@@ -62,4 +72,5 @@ def test_artifact_creation_offline() -> None:
         _artifact.delete()
     _run.delete()
     _folder.delete()
+
